pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fcra_researcher"
version = "0.1.0"
description = "Automated FCRA Compliance Researcher"
authors = [{ name = "Matrix Researcher" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "langchain>=0.2.12",
  "langchain-community>=0.2.10",
  "tiktoken>=0.7.0",
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "httpx>=0.27.0",
  "beautifulsoup4>=4.12.3",
  "lxml>=5.2.2",
  "python-dateutil>=2.9.0.post0",
  "rich>=13.7.1",
  "typer>=0.12.3",
  "sqlalchemy>=2.0.31",
  "alembic>=1.13.2",
  "sqlite-utils>=3.37",
  "redis>=5.0.7",
  "celery>=5.4.0",
  "flower>=2.0.1",
  "jinja2>=3.1.4",
  "pytest>=8.3.2",
  "pytest-mock>=3.14.0",
  "pygithub>=2.3.0",
  "requests>=2.32.3",
  "tenacity>=8.5.0",
  "openai>=1.37.0",
]

[project.optional-dependencies]
faiss = [
  "faiss-cpu>=1.8.0.post1"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app", "app.*"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["app/tests"]
